requirements.txt
openai-whisper
torch
click
pytest
pytest-cov
black
pylint>=3.0.2
PySide6>=6.6.0  # For GUI components
psutil>=5.9.0   # For system monitoring